# lib/api-zod/src/generated/types/createPrenotazioneBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePrenotazioneBody {
  vetturaId: number;
  clienteId: number;
  dataInizio: Date;
  dataFine: Date;
  stato: string;
  /** @nullable */
  note?: string | null;
}