# lib/api-zod/src/generated/types/updateVetturaBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateVetturaBody {
  /** @nullable */
  marca?: string | null;
  /** @nullable */
  modello?: string | null;
  /** @nullable */
  anno?: number | null;
  /** @nullable */
  targa?: string | null;
  /** @nullable */
  carburante?: string | null;
  /** @nullable */
  stato?: string | null;
  /** @nullable */
  colore?: string | null;
  /** @nullable */
  prezzo?: number | null;
  /** @nullable */
  km?: number | null;
  /** @nullable */
  disponibile?: boolean | null;
  /** @nullable */
  note?: string | null;
}