# lib/api-zod/src/generated/types/vetturaStorico.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestione Prenotazioni Concessionaria
 * OpenAPI spec version: 0.1.0
 */
import type { Contratto } from "./contratto";
import type { Prenotazione } from "./prenotazione";
import type { Vettura } from "./vettura";

export interface VetturaStorico {
  vettura: Vettura;
  prenotazioni: Prenotazione[];
  contratti: Contratto[];
}